guest cart works fully now. mergable upon logging in or creating new user

+++ pages/cart.php
@@ -2,23 +2,15 @@
 include('../config/essentials.php');
 include('config/edit_cart.php');
 
-if (!isLoggedIn()) {
-	// $_SESSION['msg'] = "You must log in first";
-	// header('location: login.php');
-};
-
 if (isLoggedIn()) {
 	$_SESSION['shopping_cart'] = grabUserCart();
-	$total_price = calcTotalPrice();
-	// $_SESSION['total_price'] = $total_price;
+	$total_price = calcTotalPrice($_SESSION['shopping_cart']);
 } else {
-	// $_SESSION['shopping_cart'] = array();
-	// $_SESSION['guest_cart'] = array();
-	$total_price = 0;
+	if (!isset($_SESSION['shopping_cart'])) {
+		$_SESSION['shopping_cart'] = array();
+	}
+	$total_price = calcTotalPrice($_SESSION['shopping_cart']);
 }
-
-// $_SESSION['shopping_cart'] = $cart;
-echo "<pre>" . print_r($_SESSION['shopping_cart'], true) . "</pre>";
 
 ?>
 
@@ -57,7 +49,13 @@ if (count($_SESSION["shopping_cart"]) > 0) :
 					<td><?php echo $product["title"]; ?></td>
 					<td>
 						<form method='post' action='cart.php'>
-							<input type='hidden' name='cart_id' value="<?php echo $product["cart_id"]; ?>" />
+							<?php if (isLoggedIn()) { ?>
+								<input type="hidden" name="cart_id" value="<?php echo $product["cart_id"]; ?>" />
+							<?php } else { ?>
+								<input type="hidden" name="cart_id" value="<?php echo $product["product_id"]; ?>" />
+							<?php } ?>
+
+
 							<input type='hidden' name='update_cart' value="change" />
 							<select name='quantity' onChange="this.form.submit()">
 								<option <?php if ($product["quantity"] == 1) echo "selected"; ?> value="1">1</option>
@@ -70,7 +68,13 @@ if (count($_SESSION["shopping_cart"]) > 0) :
 					</td>
 					<td><?php echo "$" . $product["price"]; ?></td>
 					<td><?php echo "$" . $product["price"] * $product["quantity"] . ".00"; ?></td>
-					<td><a href="cart.php?delete_from_cart=<?php echo $product["cart_id"]; ?>">Delete</a></td>
+
+					<?php if (isLoggedIn()) { ?>
+						<td><a href="cart.php?delete_from_cart=<?php echo $product["cart_id"]; ?>">Delete</a></td>
+					<?php } else { ?>
+						<td><a href="cart.php?delete_from_cart=<?php echo $product["product_id"]; ?>">Delete</a></td>
+					<?php } ?>
+
 				</tr>
 			<?php } ?>
 			<tr>
@@ -91,9 +95,7 @@ else :
 	<h3>Your cart is empty!</h3>
 	<p>Try adding something to cart...</p>
 <?php
-	// echo $_SESSION['guest_cart'];
 	echo 'hhhhhh';
-	echo "<pre>" . print_r($_SESSION['guest_cart'], true) . "</pre>";
 endif;
 ?>
